Replaced CustomLogging with Elasticsearch
