+++ api/.env.example
@@ -4,21 +4,21 @@ PORT=3000
 API_ACCESS_TOKEN_SECRET=oasbgasdfhdfhjtdsbsgdjzs54513afgeaer
 API_REFRESH_TOKEN_SECRET=oasbgasdfhdfhjtdsbsgdjzs54513afgeaerasdf
 
+# ElasticStack
+ELASTICSEARCH_NODE_URI=http://localhost:9200
+ELASTICSEARCH_USERNAME=elastic
+ELASTICSEARCH_PASSWORD=securePassword
+
 # Bot startup options
-DISCORD_TOKEN=YOUR_TOKEN_HERE
 DISCORD_CLIENT_ID=YOUR_CLIENT
 DISCORD_CLIENT_SECRET=YOUR_SECRET
 DISCORD_ACCESS_TOKEN_URI=https://discord.com/api/oauth2/token
 DISCORD_AUTHORIZATION_URI=https://discord.com/api/oauth2/authorize
 DISCORD_REDIRECT_URI=http://localhost:3000/auth/login/callback
-BOT_PREFIX=!
-BOT_PRESENCE_ACTIVITY_NAME=Moderation
-BOT_PRESENCE_ACTIVITY_TYPE=WATCHING
-BOT_PRESENCE_ACTIVITY_URL=https://discord.js.org
 BOT_OWNER=123456789
 BOT_INVITE=https://discord.com
 
-# Databse
+# Database
 TYPEORM_CONNECTION=postgres
 TYPEORM_HOST=localhost
 TYPEORM_USERNAME=discord
