+++ api/package.json
@@ -16,12 +16,14 @@
   "author": "Azorimor",
   "license": "GNU GPLv3",
   "dependencies": {
+    "@elastic/elasticsearch": "^7.9.0",
     "body-parser": "^1.19.0",
     "class-validator": "^0.12.2",
     "cors": "^2.8.5",
     "discord-oauth2": "^2.3.0",
     "discord.js": "^12.2.0",
     "dotenv": "^8.2.0",
+    "elastic-apm-node": "^3.7.0",
     "express": "^4.17.1",
     "helmet": "^4.1.0",
     "jsonwebtoken": "^8.5.1",
@@ -31,6 +33,7 @@
     "redis": "^3.0.2",
     "typeorm": "^0.2.25",
     "winston": "^3.3.3",
+    "winston-elasticsearch": "^0.10.0",
     "winston-transport": "^4.4.0"
   },
   "devDependencies": {
