+++ api/src/index.ts
@@ -1,6 +1,6 @@
 import app from './app';
-import config from './config';
-import logger from './logger';
+import config from './utils/config';
+import logger from './utils/logger';
 import {
   createConnection,
 } from 'typeorm';
