+++ bot/.env.example
@@ -12,3 +12,8 @@ BOT_PRESENCE_ACTIVITY_TYPE=WATCHING
 BOT_PRESENCE_ACTIVITY_URL=https://discord.js.org
 BOT_OWNER=123456789
 BOT_INVITE=https://discord.com
+
+# ElasticStack
+ELASTICSEARCH_NODE_URI=http://localhost:9200
+ELASTICSEARCH_USERNAME=elastic
+ELASTICSEARCH_PASSWORD=securePassword
