+++ bot/src/index.ts
@@ -1,12 +1,12 @@
 import {
   ShardingManager,
 } from 'discord.js';
-import config from './config';
+import logger from './utils/logger';
+import config from './utils/config';
 
-console.log("index.ts: "+config.token);
 const manager = new ShardingManager('./dist/src/instance.js', {token: config.token});
 
 manager.on('shardCreate', (shard) => {
-  console.log(`Launched shard ${shard.id}`);
+  logger.info(`Launched shard ${shard.id}`);
 });
 manager.spawn();
