+++ bot/src/utils/config.ts
@@ -11,6 +11,9 @@ const config = {
   token: process.env.DISCORD_TOKEN,
   apiAccessTokenSecret: process.env.API_ACCESS_TOKEN_SECRET,
   apiRefreshTokenSecret: process.env.API_REFRESH_TOKEN_SECRET,
+  elasticseachNodeURI: process.env.ELASTICSEARCH_NODE_URI,
+  elasticsearchUsername: process.env.ELASTICSEARCH_USERNAME,
+  elasticsearchPassword: process.env.ELASTICSEARCH_PASSWORD,
   port: process.env.PORT,
   prefix: botPrefix,
   activity_name: process.env.BOT_PRESENCE_ACTIVITY_NAME,
