+++ elk/apm.yml
@@ -2,8 +2,8 @@
 ## Default Kibana configuration from Kibana base image.
 ## https://github.com/elastic/kibana/blob/master/src/dev/build/tasks/os_packages/docker_generator/templates/kibana_yml.template.js
 #
-server.name: apm
-server.host: 0.0.0.0
+apm-server.name: apm
+apm-server.host: 0.0.0.0:8200
 elasticsearch.hosts: [ "http://elasticsearch:9200" ]
